[working] GUI 상태 정리 및 ESC 기능과 ALT+F4 기능 추가
[1]
-이전 게임 상태를 스택에 기억하도록 하여 기존에 중복되던 상태 수를 줄임. 기존에는 같은 옵션 창 상태더라도 로비 상태에서 온 것인지, 일시 중지 상태에서 온 것인지에 따라 다른 상태를 두어야 했으나, 이를 해결함.
-로비 상태, 인 게임 상태는 각각 중심 상태로 여겨 해당 상태에 있을 때는 스택이 비워져있도록 설계함.
-이를 위해 게임으로의 로딩이나, 로비로의 로딩 과정에서도 스택을 비우도록 함.

[2]
-로비 상태나 인 게임 상태와 같은 중심 상태로 돌아가는 방향으로의 버튼 클릭시 발동되던 조건문들은 ESC 키로도 활성화되도록 단축키 기능을 구현함.
-GetKeyState를 사용할까 고민했지만, ImGui가 제공하는 키 눌림 여부 반환 함수 ImGui::IsKeyPressed(..)가 있어 그것을 사용함. GUI 업데이트 메소드에서 해당 여부를 확인하는 것이므로, ImGui 함수를 이용하는 것이 더 적합하다고 판단함. 해당 함수를 사용할 때 두 번째 매개변수엔 false를 두어서 눌리는 시점 한 번만 조건문이 발동되도록 하여 한 번의 클릭으로 연달아 상태가 변화되는 일이 없도록 함.

[3]
-ALT+F4로 게임을 종료할 수 있도록 단축키 기능을 제공함.
-인 게임 상태, 일시 정지 상태, 일시 정지를 거친 옵션 조절 상태인 경우엔 종료되면 세이브되지 못한 것들이 소실될 수 있으므로 확인창을 띄우는 상태로 가도록 하였음.
-로비 상태나 로비를 거친 옵션 조절 상태나, 로딩 상태는 종료를 하여도 소실될 것이 없으므로 ALT+F4가 눌리면 바로 종료되도록 하였음.
-종료 확인 상태에서 ALT+F4를 누른다는 것은 종료하겠다는 의사가 있다는 것으로 판단하여 바로 종료되도록 하였음. 즉 어떤 상태에 있든간에 ALT+F4를 두 번 연속해서 누른다면 바로 종료되도록 함.

[4]
-ImGui의 버튼이나 공백 크기들의 베이스 크기를 저장하던 상수 정적 멤버 변수들을, 스케일을 반영하며 업데이트되는 일반 멤버 변수로 수정하였음.
-기존엔 베이스 크기를 명확히 남겨둬야한다는 생각해 상수로 두었는데, 생각해보니 베이스크기는 스케일을 곱하는 식의 상수로 남겨둘 수 있으니 굳이 따로 보존할 필요가 없었음.
-이를 통해 GUI 업데이트 함수에서 베이스 크기에 스케일을 곱하여 사용하던 중복된 구문들을 줄였음.

[5]
-SodoUtil.cpp에 두던 FindMostIntersectingOutput(..) 메소드는 유틸리티라기보다는 초기화 과정에 속하는 메소드라서 위치를 옮김.
-그리고 CheckAndExit()은 그냥 둘 필요 없이 본문을 WM_CLOSE 메시지 처리부에 작성하는 것이 더 깔끔하기에 삭제함.
-이로 인해 SodoUtil.cpp엔 둘 내용이 없어졌으므로 파일을 삭제함.

+++ Code/Source/SodoReset.cpp
@@ -93,7 +93,7 @@ void Sodo::ResetScreenSetting()
 	{
 		ResetQueuedCommands();
 
-		for (int i = 0; i < m_screenBackBufferCount; i++)
+		for (UINT i = 0; i < m_screenBackBufferCount; i++)
 		{
 			m_screenBackBuffers[i].Reset();
 		}
@@ -135,10 +135,10 @@ void Sodo::ResetScreenSetting()
 
 void Sodo::ResetToFullScreenMode()
 {
-	INT tempPreviousWindowPosX = 0;
-	INT tempPreviousWindowPosY = 0;
-	INT tempPreviousWindowWidth = 0;
-	INT tempPreviousWindowHeight = 0;
+	LONG tempPreviousWindowPosX = 0;
+	LONG tempPreviousWindowPosY = 0;
+	LONG tempPreviousWindowWidth = 0;
+	LONG tempPreviousWindowHeight = 0;
 	RECT rectWindow = {};
 	BOOL queryResult = GetWindowRect(m_hWnd, &rectWindow);
 	if (queryResult != FALSE)
